+++ certbot/main.py
@@ -14,7 +14,7 @@ from acme import errors as acme_errors
 from acme.magic_typing import Union  # pylint: disable=unused-import, no-name-in-module
 
 import certbot
-from certbot import account
+from certbot._internal import account
 from certbot import cert_manager
 from certbot import cli
 from certbot import client
@@ -483,7 +483,7 @@ def _determine_account(config):
 
     :returns: Account and optionally ACME client API (biproduct of new
         registration).
-    :rtype: tuple of :class:`certbot.account.Account` and :class:`acme.client.Client`
+    :rtype: tuple of :class:`certbot._internal.account.Account` and :class:`acme.client.Client`
 
     :raises errors.Error: If unable to register an account with ACME server
