+++ certbot/tests/account_test.py
@@ -1,4 +1,4 @@
-"""Tests for certbot.account."""
+"""Tests for certbot._internal.account."""
 import datetime
 import json
 import unittest
@@ -19,10 +19,10 @@ KEY = jose.JWKRSA.load(test_util.load_vector("rsa512_key.pem"))
 
 
 class AccountTest(unittest.TestCase):
-    """Tests for certbot.account.Account."""
+    """Tests for certbot._internal.account.Account."""
 
     def setUp(self):
-        from certbot.account import Account
+        from certbot._internal.account import Account
         self.regr = mock.MagicMock()
         self.meta = Account.Meta(
             creation_host="test.certbot.org",
@@ -31,9 +31,9 @@ class AccountTest(unittest.TestCase):
         self.acc = Account(self.regr, KEY, self.meta)
         self.regr.__repr__ = mock.MagicMock(return_value="i_am_a_regr")
 
-        with mock.patch("certbot.account.socket") as mock_socket:
+        with mock.patch("certbot._internal.account.socket") as mock_socket:
             mock_socket.getfqdn.return_value = "test.certbot.org"
-            with mock.patch("certbot.account.datetime") as mock_dt:
+            with mock.patch("certbot._internal.account.datetime") as mock_dt:
                 mock_dt.datetime.now.return_value = self.meta.creation_dt
                 self.acc_no_meta = Account(self.regr, KEY)
 
@@ -55,18 +55,18 @@ class AccountTest(unittest.TestCase):
           "<Account(i_am_a_regr, 7adac10320f585ddf118429c0c4af2cd, Meta("))
 
 class ReportNewAccountTest(test_util.ConfigTestCase):
-    """Tests for certbot.account.report_new_account."""
+    """Tests for certbot._internal.account.report_new_account."""
 
     def _call(self):
-        from certbot.account import report_new_account
+        from certbot._internal.account import report_new_account
         report_new_account(self.config)
 
-    @mock.patch("certbot.account.zope.component.queryUtility")
+    @mock.patch("certbot._internal.account.zope.component.queryUtility")
     def test_no_reporter(self, mock_zope):
         mock_zope.return_value = None
         self._call()
 
-    @mock.patch("certbot.account.zope.component.queryUtility")
+    @mock.patch("certbot._internal.account.zope.component.queryUtility")
     def test_it(self, mock_zope):
         self._call()
         call_list = mock_zope().add_message.call_args_list
@@ -74,10 +74,10 @@ class ReportNewAccountTest(test_util.ConfigTestCase):
 
 
 class AccountMemoryStorageTest(unittest.TestCase):
-    """Tests for certbot.account.AccountMemoryStorage."""
+    """Tests for certbot._internal.account.AccountMemoryStorage."""
 
     def setUp(self):
-        from certbot.account import AccountMemoryStorage
+        from certbot._internal.account import AccountMemoryStorage
         self.storage = AccountMemoryStorage()
 
     def test_it(self):
@@ -92,16 +92,16 @@ class AccountMemoryStorageTest(unittest.TestCase):
 
 
 class AccountFileStorageTest(test_util.ConfigTestCase):
-    """Tests for certbot.account.AccountFileStorage."""
+    """Tests for certbot._internal.account.AccountFileStorage."""
     #pylint: disable=too-many-public-methods
 
     def setUp(self):
         super(AccountFileStorageTest, self).setUp()
 
-        from certbot.account import AccountFileStorage
+        from certbot._internal.account import AccountFileStorage
         self.storage = AccountFileStorage(self.config)
 
-        from certbot.account import Account
+        from certbot._internal.account import Account
         new_authzr_uri = "hi"
         self.acc = Account(
             regr=messages.RegistrationResource(
@@ -162,7 +162,7 @@ class AccountFileStorageTest(test_util.ConfigTestCase):
         # pylint: disable=protected-access
         self.storage._load_for_server_path = mock.MagicMock(
             side_effect=["x", errors.AccountStorageError, "z"])
-        with mock.patch("certbot.account.os.listdir") as mock_listdir:
+        with mock.patch("certbot._internal.account.os.listdir") as mock_listdir:
             mock_listdir.return_value = ["x", "y", "z"]
             self.assertEqual(["x", "z"], self.storage.find_all())
 
@@ -171,7 +171,7 @@ class AccountFileStorageTest(test_util.ConfigTestCase):
 
     def _set_server(self, server):
         self.config.server = server
-        from certbot.account import AccountFileStorage
+        from certbot._internal.account import AccountFileStorage
         self.storage = AccountFileStorage(self.config)
 
     def test_find_all_neither_exists(self):
