--- docs/api/account.rst
@@ -1,5 +1,0 @@
-:mod:`certbot.account`
---------------------------
-
-.. automodule:: certbot.account
-   :members:
